+++ packages/soeditor/src/vite-env.d.ts
@@ -1,6 +1,7 @@
 interface ImportMeta {
     readonly env: {
         readonly SOEDITOR_OPTIONAL_CLASSIC?: string;
+        readonly SOEDITOR_STANDALONE_VIDEO?: string;
         readonly SOEDITOR_TABLE_CONTEXT?: string;
     };
 }